IMPORTANT fix: middleware previously used UserWithoutPassword to User typecast, which is unsafe.
