+++ backend/user-service/src/middleware/authMiddleware.ts
@@ -3,15 +3,21 @@ import { authenticateAccessToken } from "../utils/jwt";
 
 export interface User {
   id: number;
+  password: string;
   username: string;
   email: string;
   role: string;
-  languages: string[];
-  token: string;
+  languages: { id: number; language: string }[];
+  token?: string;
 }
 
-interface JwtPayload {
-  user: User;
+export interface UserWithoutPassword {
+  id: number;
+  role: string;
+}
+
+export interface JwtPayload {
+  user: UserWithoutPassword;
   exp: number;
   iat: number;
 }
@@ -30,11 +36,7 @@ export async function verifyAccessToken(
       .json({ errors: [{ msg: "Not authorized, no access token" }] });
   } else {
     try {
-      const decoded = (await authenticateAccessToken(
-        accessToken,
-      )) as JwtPayload;
-      req.user = decoded.user;
-
+      await authenticateAccessToken(accessToken);
       next();
     } catch (error) {
       res
